Ported Schema and Credential Registry APIs from Mark I agent.
Hit a snag with Registries (and all TEL operations) that we will need a refactor in KERIpy to get them to support Signify supplied signed events, like we did for Hab / Habery.  Pushing this now with all Endpoints working except the final step of registry creation.

Added test coverage for agenting.loadEnds too